Move observable into it's own file
